MORE-978: Added Bluetooth Setup Screen in Android and fixed iOS loading tasks after a relogin

+++ iosApp/iosApp/Services/DataUploadManager.swift
@@ -24,24 +24,24 @@ class DataUploadManager {
                         let userDefaults = UserDefaultsRepository()
                         let networkService = NetworkService(endpointRepository: EndpointRepository(sharedStorageRepository: userDefaults), credentialRepository: CredentialRepository(sharedStorageRepository: userDefaults))
                         print("Sending data to backend...")
-                        
-                        networkService.sendData(data: dataBulk) { [weak self] pair in
-                            if let error = pair.second {
-                                print("Error: \(error)")
-                                self?.currentlyUploading = false
-                                completion(false)
-                            } else if let self, let idSet = pair.first as? Set<String> {
-                                print("Sent data! Deleting local data...")
-                                self.observationDataRepository.deleteAllWithId(idSet: idSet)
-                                print("Deleted data!")
-                                self.currentlyUploading = false
-                                completion(true)
-                            } else {
-                                print("Error!")
-                                self?.currentlyUploading = false
-                                completion(false)
+                        networkService.iosSendData(data: dataBulk) { weakReference in
+                            if let pair = weakReference.value {
+                                if let error = pair.second {
+                                    print("Error: \(error)")
+                                    self?.currentlyUploading = false
+                                    completion(false)
+                                } else if let self, let idSet = pair.first as? Set<String> {
+                                    print("Sent data! Deleting local data...")
+                                    self.observationDataRepository.deleteAllWithId(idSet: idSet)
+                                    print("Deleted data!")
+                                    self.currentlyUploading = false
+                                    completion(true)
+                                } else {
+                                    print("Error!")
+                                    self?.currentlyUploading = false
+                                    completion(false)
+                                }
                             }
-                            networkService.close()
                         }
                     } else {
                         print("No data to send!")
@@ -54,8 +54,5 @@ class DataUploadManager {
     }
     func close() {
         print("Closed!")
-    }
-    deinit {
-        
     }
 }
